Time: 13 ms (68.30%), Space: 11.2 MB (6.25%) - LeetHub

+++ 462-minimum-moves-to-equal-array-elements-ii/462-minimum-moves-to-equal-array-elements-ii.cpp
@@ -3,7 +3,7 @@ public:
     int minMoves2(vector<int>& nums) {
         int n=nums.size();
         sort(nums.begin(), nums.end());
-        vector<long long> prefix(nums.begin(), nums.end());
+        vector<int> prefix(nums.begin(), nums.end());
         for(int i=1;i<n; i++){
             prefix[i]+=prefix[i-1];
         }
